Adding madge circular dependency check and fixing circular dep issues!
