+++ library/data-model/package.json
@@ -35,7 +35,8 @@
     "test": "NODE_ENV=test jest",
     "test:watch": "NODE_ENV=test jest --watchAll=true",
     "coverage": "NODE_ENV=test jest --coverage --watchAll=false",
-    "lint": "gts lint",
+    "circular-dep-check" : "madge -c --extensions ts src/",
+    "lint": "npm run circular-dep-check && gts lint",
     "fix": "gts fix"
   },
   "author": "",
@@ -65,6 +66,7 @@
     "fast-check": "^3.6.3",
     "gts": "^5.3.1",
     "jest": "^29.4.3",
+    "madge": "^8.0.0",
     "pouchdb-adapter-memory": "^8.0.1",
     "pouchdb-find": "^8.0.1",
     "ts-jest": "^29.0.5",
