+++ library/data-model/src/constants.ts
@@ -1,0 +1,5 @@
+export const ENCODING_SEPARATOR = '||';
+export const COUCH_ADMIN_ROLE_NAME = '_admin';
+
+// This is configurable, but it's fine to use the default
+export const COUCHDB_ROLES_PATH = '_couchdb.roles';
