+++ library/data-model/src/data_storage/authDB/security.ts
@@ -2,8 +2,8 @@
  * This module exports the security document to be used for the auth database.
  */
 
+import {COUCH_ADMIN_ROLE_NAME} from '../../constants';
 import {SecurityDocument} from '../utils';
-import {COUCH_ADMIN_ROLE_NAME} from '../..';
 
 export const AuthDatabaseSecurityDocument: SecurityDocument = {
   // Only DB admins or cluster admins can interact with the AuthDB
