+++ library/data-model/src/data_storage/internals.ts
@@ -19,7 +19,6 @@
  */
 
 import {v4 as uuidv4} from 'uuid';
-
 import {HRID_STRING} from '../datamodel/core';
 import {
   getAttachmentDumperForType,
@@ -31,7 +30,7 @@ import {
   getHridFieldNameForViewset,
   getIdsByFieldName,
   HridFieldMap,
-} from '../index';
+} from '../utils';
 import {
   Annotations,
   AttributeValuePair,
