+++ library/data-model/src/data_storage/invitesDB/security.ts
@@ -2,7 +2,7 @@
  * This module exports the security document to be used for the auth database.
  */
 
-import {COUCH_ADMIN_ROLE_NAME} from '../..';
+import {COUCH_ADMIN_ROLE_NAME} from '../../constants';
 import {SecurityDocument} from '../utils';
 
 export const InvitesDBSecurityDocument: SecurityDocument = {
