+++ library/data-model/src/data_storage/merging.ts
@@ -20,8 +20,7 @@
  *   merging by humans.
  */
 
-import {getDataDB} from '../index';
-
+import {getDataDB} from '../callbacks';
 import {
   AttributeValuePairIDMap,
   DataDbType,
