+++ library/data-model/src/data_storage/projectsDB/types.ts
@@ -1,4 +1,4 @@
-import {PossibleConnectionInfo} from '../..';
+import {PossibleConnectionInfo} from '../../types';
 
 // V1
 export type ProjectV1Fields = {
