+++ library/data-model/src/datamodel/core.ts
@@ -22,7 +22,7 @@
  */
 
 import {
-    CouchProjectUIModel,
+  CouchProjectUIModel,
   EncodedProjectUIModel,
   FullyResolvedRecordID,
   ProjectUIModel,
@@ -69,14 +69,28 @@ export function split_full_record_id(
  * @returns Normal functional spec
  */
 export const decodeUiSpec = (
-  rawUiSpec: CouchProjectUIModel
+  rawUiSpec: EncodedProjectUIModel
 ): ProjectUIModel => {
   return {
-    _id: rawUiSpec._id,
-    _rev: rawUiSpec._rev,
     fields: rawUiSpec.fields,
     views: rawUiSpec.fviews,
     viewsets: rawUiSpec.viewsets,
     visible_types: rawUiSpec.visible_types,
+  };
+};
+
+/**
+ * converts the encoded UI spec representation into the normal one
+ * @param rawUiSpec The raw encoded spec
+ * @returns Normal functional spec
+ */
+export const decodeCouchUiSpec = (
+  rawUiSpec: CouchProjectUIModel
+): ProjectUIModel => {
+  const {_id, _rev, ...spec} = rawUiSpec;
+  return {
+    _id,
+    _rev,
+    ...decodeUiSpec(spec),
   };
 };
