+++ library/data-model/src/index.ts
@@ -25,6 +25,7 @@ export * from './callbacks';
 export * from './logging';
 export * from './types';
 export * from './utils';
+export * from './constants';
 
 // Nested folders
 export * from './datamodel';
