+++ library/data-model/src/permission/index.ts
@@ -2,3 +2,4 @@ export * from './model';
 export * from './functions';
 export * from './tokenEncoding';
 export * from './userFunctions';
+export * from './types';
